Refactors package into one simpler module
