+++ pyTrnsysType/__init__.py
@@ -8,4 +8,3 @@ warn_if_outdated("pyTrnsysType", __version__)
 
 from .utils import *
 from .trnsymodel import *
-from .input_file import *
